add:hook_NewForm
