+++ frontend/src/components/templates/Productnew.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react';
 import { Header } from 'semantic-ui-react';
-import Newform from 'components/organisms/Newform';
+import Newform from 'container/EnhancedNewForm';
 
 const ProductNew: FC = () => (
   <>
